// src/app/layouts/economie/services/models/economie.ts
export class Economie {
  id?: string;
  TauxDeCHangeDollard?: string;
  TauxDeCHangeFc?: string;

  // Prix de denrée de premier neccesiteux
  PrixDenreMais?: string;
  PrixDenreRiz?: string;
  PrixDenreHaricot?: string;
  PrixDenreSucre?: string;
  PrixDenreLait?: string;
  PrixDenreThe?: string;
  PrixDenreCafe?: string;
  PrixDenreTomate?: string;
  PrixDenreSel?: string;
  PrixDenreAil?: string;
  PrixDenreOignon?: string;
  PrixDenreSavon?: string;
  PrixDenreMargarin?: string;
  PrixDenreSpaghetti?: string;
  PrixDenreFroma?: string;
  PrixDenrePoisson?: string;
  PrixDenreHuileRouge?: string;
  PrixDenreHuileJaune?: string;

  // Prix Du Carburant
  PrixDuEssence?: string;
  PrixDuPretrole?: string;
  PrixDuMazoute?: string;

  QualiteRN?: string;
  NomRN?: string;
  QualiteCheminDeFer?: string;
  QualiteAeroport?: string;
  NomAeroport?: string;

  RapportGeneralEconomique?: string;

  // Agent?: string;
  // Nation?: string;
  Province?: string;
  // Territoire?: string;
  CommunaleUrbain?: string;
  CheferieSecteur?: string;

  Created?: any;
  agentId?: string;
  // ecoImageUrl: string;
}